requirements.txt
colorlog==6.6.0
Flask==2.0.2
pycryptodome==3.14.1
Werkzeug==2.0.2
requests==2.26.0
PyMySQL~=1.0.2
PyJWT~=2.7.0